ESTCj: add a --model option so that we may use libgas or cea2 gases.
The old behaviour of using cea2 gases is default so command-line users should
be able to run exactly as they have before.

The make_gas_from_name function in libgas_gas can now accept a filename for
a Lua gas model file (which can also be a compressed look-up table).