Eilmer3 house-keeping: appease the clang 3.4 compiler.
There were a few unused variables and one use of an uninitialized variable
(theta_A0_ in species_energy_modes.cxx) that really was a bug, now fixed.
Most changes just identified the unused variables in the code directly.
In a few places, names of members were changed so that the formal arguments
of the constructor did not shadow them.