Eilmer3: added MassFluxOutBC for Anand and Ingo.
I think that the code is complete, however, the actual BC hasn't been tested.
The automated tests are running happily as I make this commit, so I think that
it hasn't broken anything extra but I do recall that nenzfr was broken in
yesterday's test run :-(  Fixing that is a job for the next commit.