nenzfr: fixed the nozzle template file to pass the cea2 module to rstc.
A couple of weeks ago, the ESTCj was generalized a little to work on more gas
models than just the cea2-backed model.  There's now the libgas-backed model
as well.  nenzfr, via the nozzle template file, had continued to assume that
the cea2_gas model was being used.  That is now more explicit with this commit.
nenzfr should also be allowed to work with the libgas-based model but that's
a job for another day.