Eilmer3: finish implementation of MassFluxOutBC and adjust SubsonicInBC.
The adjustment to SubsonicInBC is to use the locally-inward direction
as the direction for the computed velocity for the ghost cell.
There was also a bit of house-cleaning of the code.  Whether it's tidier
probably depends on the person looking at it ;-)

Also, provide an example of this pair of boundaries establishing a slow
(subsonic) flow in a tube.