app/nenzfr: a few small changes to nenzfr
it turns out most users won't be in the group that the run_template_Barrine.sh
was using, so I changed it to MechMiningEng. Changed nenzfr_utils.py as I
hadn't told it to add one of the variables to the substitute dictionary
for the run template.
Also made a change to the nenzfr perturbed input checkers so it knew
how to set the default run command for Barrine as well as Mango.