Eilmer3: fixed bug identified by Dan; coupling of radiation to flow updates.
Since the radiation source term is computed only once before starting the staged
update of the gas dynamics, we need to not throw its value away on clearing the
source vector (or, recover the value after clearing the source vector).
This commit has taken the option of restoring the source term because I though
that was the easiest to maintain fix.