lib/nm: Reinstated Valmatrix::gaussian_elimination
The standalone function was not thread safe, whereas this implementation
inside the Valmatrix class is.  This is necessary for running OpenMP
e3rad calculations when the QSS modelling is turned on, as it uses
Gaussian elimination to solve the system of equations.

Perhaps the standalone function is not necessary since the operation
must always be performed on a Valmatrix instance, and the row_index
array can be stored as a member of the class.