Added example for quick calculation of X2 VUV air spectra.
This example calculates X2 air spectra and plots in the VUV
if your 'equilibrium' region mass fractions, temperature and
pressure are known from CFD. The data comes from Umar's
condition 1 simulations (see his upcoming thesis) for a flat
faced model in X2 with air as the test gas, and could be
an easy way to generate spectra from CFD and compare to
experiment. Parade is an option if you have access.