onedval: Removed the dependency on e3prep.py and libprep3.py
By using the gaspy module and the minimal_geometry.py module directly,
I was able to remove the dependency on eilmer. I should have done
this a while ago. It was just laziness at the time that lead me to use
the eilmer modules directly. Now onedval does NOT require a full install
of eilmer to function.