eilmer3: Reinstated -std=c++0x option for CXXFLAG3
I'd accidentally committed over this change at some point.
Frustratingly, I require this flag to be present for compilation
on raijin BUT it should NOT appear when I compiler on my
ScientificLinux box using the GNU 4.8 compiler. A better fix
is to pass this through as an option.... for another day.