lib/radiation: improved the new atomic radiator input data
1. The NIST-ASD data for O+ was actually for N+; this has now been fixed
2. Previously multiple PICS could be assigned to the same electronic level; this is now limited to the closest one by energy
3. Allowed a few more parameters such as "allow_inexact_Stark_matches" to be set in the users python script

The comparison between the legacy and new input data sets in examples/photaura/equilibrium-spectra/EAST/N2-O2/shot_46-45/
now runs, but the comparison is not too good.  This is largely because only a small number of the available photoionization
cross-sections are making their way into the c++ code due to many of them having the same level index.  The TOPBase PICS c++
classes may have to be changed to allow multiple vectors of nu and sigma.