Updated Ramshaw-Chang multicomponent diffusion model.
The Ramshaw-Chang model has been updated and improved by my colleagues
at EPFL, so now it is fully functional for our aeroshell-type
simulations. It is slightly more computationally expensive than the
Constant Lewis number model we were using previously, but we hope that
the calculation of diffusion coefficients (rather than having constant
values) will improve results. A test case has been constructed, and
once I tidy it up, it will be placed in the eilmer3 examples.

Also, the Lewis number value was changed from 1.4 to 1.0 to be in line
with the test case.