eilmer3: More information now passed to timestep start/end Lua hooks.
The global Lua environment handed to the at_timestep_start()
and at_timestep_end() functions (as well as the source_vector()
function) now includes:
nblks  -- number of blocks active on local rank
blks   -- list of block infor on local rank

I think it will work to just proved information on
a per rank basis because each block has a globally
unique id and the user can access that in the 'blks'
information.

The documentation has been updated and an example
added to the cone20-udf case.