Eilmer3: Add capability to monitor temperature at user-specified points...
and to be able to stop if any point sees a temperature change greater than
a specified tolerance.

This is mainly to make it more convenient to do coupled-radiation calculations.
We don't want the flow field changing too much between radiation calculations
else the whole simulation seems to go unstable.

There is an example of setting some points in the 2D/sod/N2-O2/sod-monitor.py
where the simulation is terminated early, as the shock crosses a monitor point.