Eilmer3: Make symmetric the propagation of the cell lengths into the ghost cells.
Jason Qin and Paul Petrie-Repar have identified the lack of exact symmetry in
the reconstruction process at the wall as being a cause of the leaky wall
boundary conditions.  Note that the symmetry is not consistent with the
linear extrapolation used for the positions and volumes in the following section
of block_geometry code (which Rowan added to get high-order extrapolation for
boundary conditions, if I recall correctly).  We think about this carefully.