shock_layer_surface.py: some small modifications to shock fitting functions
1) Just vary bx_scale and by_scale when fitting a Billig curve to a shock.
   From a number of tests it seems that there is no great benefit in allowing
   Mach number and radius to vary.

2) Allow user-definable weights to be given to the fit_spline2shock function.

Currently these shock fitting functions are not working 100% as I would like.
It is hard to get a curve to be well behaved both near the stagnation streamline
and towards the back of the body.  It would be nice to be able to use the shock fitting
inflow boundary condition but for the case of free electron nonequilibrium sometimes the
electron shock can procede the heavy particle shock and therefore the shock must be captured.