Eilmer3: generalized the SubsonicInBC for turbomachine simulations.
As well as the normal direction at the boundary, we have reinstated the option to
specify a vector direction and also to specift flow angles for swirling flows.
Most of the house-keeping code is in place, however, only the WEST boundary has
the full set of options coded and it really hasn't been tested in anger.
More tomorrow.