lib/gas/kinetics: Allow user to set limit on nonequilibrium factor in Knab model.
At certain gas conditions, the nonequilibrium factor (which enhances
the reaction rate) takes on silly values when using the Knab model.
This can occur when the local vibrational temperature is much higher
than the translational temperature. In these cases, the user may specify
a limit on the nonequibrlium factor using the Z_limit variable. This
will limit the amount by which the reaction rate is enhanced.