A tcl test script for Poshax3.
Currently only the Fire II case is included.  The idea of this
script is to include some quick Poshax3 calculations that test
the chemical kinetic models, and to a lesser extent the functionality
of Poshax3, that developers can run to ensure they have not adversely altered
the code from its current functionality.  Poshax3 is a good platform
to use for this as it enables the models relating to thermochemical
relaxation to be tested very quickly.