examples/poshax3: added Glass and Liu Argon cases to the tcl test script
The calculation without radiation only takes a few seconds and is included
in the short tests, while the calculation with radiation takes 10 minutes or
so to run and is therefore included in the long tests.

In this commit, I have also updated the png files from the Fire II and Glass-Liu
examples to match the output from the current version of the repository.  Also,
the spectrally_resolved flag for the radiation model is now expected to be a boolean
everywhere (both input scripts and c++ code).