examples/eilmer3: an updated version of the Rutowski hemisphere example (but still a work in progress)
Currently the whole simulation (inviscid, viscous, coupled radiation) takes 8 hours or so on 4 CPU's.
This could be reduced to 4 hours if radiation coupling proves to be unimportant.
It will be a good test case to ensure the code is performing as expected for radiation, nonequilibrium
plasmas as it takes a relatively short amount of time and has experimental data available for comparison.