eilmer3: initialise vectors of radiative heat flux at walls to zero          when starting a ray-tracing based radiation transport calculation.
Previously when an existing solution for q_rad existed, the new solution
was added on to the old result instead of replacing it. This only applies
to the ray-tracing based radiation transport models (Discrete Transfer
and Monte Carlo).

Users that have run simulations with radiation-flowfield coupling with
one of the ray-tracing models and the surface energy balance BC should
rerun with this new version of the code, as the surface temperatures
will have been overpredicted in the old version of the code.