A test case for nonequilibrium radiating shock layers in eilmer3.
The test attempts to reproduce the measured irradiance at the stagnation
point of a 1 inch sphere placed in a shock tube with an argon test gas
with the primary shock at Mach 12.70.  The experimental results were
reported by Rutowski and Bershader (Phys. Fluids. 7(4) 1964 pp 567-577).

Features of the test include:
 - separate translational temperatures for electrons and heavy particles
 - viscous boundary layer included
 - Monte-Carlo radiation transport
 - radiation-flowfield coupling (in long version of test)

Currently eilmer3 reproduces the measured irradiance (~55 MW/m2) to within
the experimental uncertainty of 16%.

The short version of the test, which doesn't consider radiation-flowfield
coupling, takes 2 hours to run on 4 cores.  The long version takes
approximately 5 hours.  The former has been added to the list of long tests
to run in eilmer3-test.tcl.