lib/util: Updated systems.mk so that OpenMPI compiles nicely on Ubuntu 12.04
In this commit, I have added Zac Denman's changes to the systems.mk file
to get the open_mpi target working on Ubuntu. OpenMPI provide wrappers for their
compilers with:
mpicc -- for C compilation; and
mpiCC -- for C++ compilation
which is what we had. However, they give the warning that mpiCC should
not be used on systems that are NOT case sensitive. It's not clear to me
why Ubuntu 12.04 would not have mpiCC available, but it apparently
did not work.

The solution is to mpicxx as the C++ compiler wrapper which is also
provided by OpenMPI installations.