Eilmer: Added a user-defined energy flux b.c.
This boundary condition is similar to the user-defined
mass flux bounday condition. In this case, the user
gets the momentum and mass effect of a no-slip wall
but is free to set the energy flux at the boundary. I
have tested the simple case of setting 0 energy flux
(adiabatic) at the wall by using Jason's Couette flow
test case.