Eilmer3: Try out the address sanitizer that is available in gcc 4.8.x.
You can invoke the build for openmpi with the script make_for_gnu48_debug.sh,
however, I have not got it to successfully finish the build on my LinuxMint-13
system with gcc 4.8.2 freshly installed into /usr/local/
This commit still puts the scripts into the repository so that we can try the
build on a more recent system.

The usual build target for_openmpi still works and runs the automated tests.