lib/util: Added TARGET=for_openmpi_intel for use on raijin.
On raijin, the recommended MPI system is their compilation of
OpenMPI using the Intel compiler. This caused some problems
using the SWIG shared objects as Python modules because some
of the pointers to dynamic intel libraries were not present.
This new TARGET in the systems.mk file fixes that by giving
some extra link directives to the Intel specific libraries.