Eilmer3: Added an IgnitionZone object that uses an effective temperature for chemistry updates.
This commit is on behalf of Xin Kang. Xin implemented an IgnitionZone
that works much like a HeatZone. The IgnitionZone applies an effective
ignition temperature when computing chemistry updates. The user sets
this effective ignition temperature. The effective temperature only
influences the rate of species change calculation. It does not alter
the gas temperature (the static temperature of the gas).

The application of this is for the inlets to microcombustors where
we'd like to kick-start some ignition as the pre-mixed fuel enters
the combustor.