Eilmer3/examples: Update mms case to use sympy for generating the source terms.
In switching from maxima to sympy, the L2 norm in density for case 2
differed in the 4th significant figure. I am happy to accept the
new value and this is reflected in the change to the mms.test script.
This difference is due to floating point representation of the constants
in the analytic source terms. These differences are plotted in the
newly added PNG files for each of the source terms.