Improving the eilmer3-PARADE coupling method.
Infrastructure to allow a radiation tranport calculation in
e3rad to read in emission and absorption coefficients that
have already been calculated in PARADE. Previously, the
generation of PARADE input files and the PARADE calculation
was done within the e3rad calculation, but this was far too
time consuming for more complex simulations.

Also, radmodel.py and parade_radiator_library.py have been
updated to improve the automatically generated parade.con
files (in line with PARADE v3.1), e.g. designation of
temperature indices, and indicating that results should be
generated for every cell.

Coming soon - python and shell scripts to generate the input
files from an eilmer3 solution, and run the PARADE calculation.