onedval: Added option to take axisymmetric data.
Bianca Capra requested that onedval process one-d averages
for axisymmetric data. To do this, a line of data is extracted
across the domain. When the 'geometry_type' is set to 'axi',
onedval will weight this data appropriately so that the
axisymmetric nature of the data is included in the average.