Eilmer3: implemented an nonuniform_t boundary condition.
This BC is used to set the nonuniform wall temperature, for top and bottom BC,
the input tempearture is written as j direction first, then i direction. for west
and east BC, j direction first, then k direction, and i direction first, then
k direction for north and south boundary.

The input parameters are T_non, which is nonuniform wall temperature,
starting_blk which is the starting block id for SuperBlock(), no_blk is the number
of blocks in each direction.

The moving velocity of wall can also be set like moving_wall BC.