lib/nm: Fixed bug in alpha-QSS error criterion.
While working with Kyle Damm on a GPU implementation
of the alpha-QSS ODE integrator, I relooked critically
at the alpha-QSS code in lib/nm. I found a bug in how the error
criterion was applied. This is now fixed. I have compared
it to the source code reported in:

Mott, D. and Oran, E. (2001)
CHEMEQ2: A Solver for the Stiff Ordinary Differential
Equations of Chemical Kinetics,
Naval Research Laboratory, Report no. NRL/MR/6400-01-8553