lib/gas: Bug fix for initialissation of SSH relaxation time models.
There was a bug in setting up the Lennard-Jones parameters
for like and unlike molecules (ie. polar and nonpolar). The
if-else logic was backwards.

TODO: The xi_correction calculation seems to be in error.
It could be the units for dipole moment. I need to check
this expression against some sources in the literature.