eilmer3: Added check volumes at prep stage in 2D.
There is now a check_volumes() method for the Block2D
Python object that will check for negative volumes
during prepartion if called by the user. The method
returns a list of tuples of form (i, j, vol) of
only those cells with zero or negative volume.

At the moment to run this check, I include this text in my Python
script after setting up all of the blocks:

for blk in Block.blockList:
    print "Checking block: ", blk.blkId
    bad_cells = blk.check_volumes(gdata)
    if bad_cells:
        for c in bad_cells:
            print "Bad cell: ", c[0], c[1], c[2]

This could be done automatically once we decide
the appopriate point in time to do it, and
have the option to switch the check on or off.