nenzfr/nozzle.input.template - bug fixes...
PJ noted two bugs that I missed when trying to run the Mach 4
equilibrium air test case. The first bug related to incorrectly
making the gasName upper-case when specifying the cea-lut
filename.

The second bug was because the massf parameter in the FlowCondition
object CANNOT be a dictionary when a LUT has been specified via.
fname in select_gas_model(). When a LUT is used, massf must be
set to either None or [1.0,].