Eilmer3: Added the TransientProfileBC class.
This is part 2 of 2 for making multi-part transient flow simulations,
however, the current implementation of the new class is incomplete
since it reads only the first time instant.  The present code is intended
to be form a basis for discussion.  There will be another commit when the
implementation is complete.