lib/gas: Fixed xi_correction calculation in SSH VT/VV relaxation times.
In this commit, I have finally had the chance to address the TODO
I gave myself in Revision 1779 on 30 Sep 2014. I found the error
in my xi_correction calculation. It turns out that the eps value
for potential well (one of the Lennard-Jones force constants) needs
to be used in units of K. In other words, the value is actually
eps/k (where k is the Boltzmann constant) in the expression for xi.

Note, I cannot actually find the provenance of this expression for
the combining law for polar-nonpolar molecules. I have found it in
two references (Hirschfelder et al [1954] and Kee et al. [1998],
of Chemkin fame). Neither of these references give a source for the
original expression, nor do they show a derivation. Furthermore,
both references have the error with the factor of 1/k missing.