Eilmer3: Added coaxial_jets case by Wilson Y. K. Chan
Added the files to replicate the coaxial_jets case developed
by Wilson Y. K. Chan for the k-omega turbulence model report.
IMPORTANT: the case is unable to be executed when utilising
the k-omega turbulence model since it was used last (2009).
All efforts to remedy this have resulted in the case crashing
(values tending to nan) before the first step.
It is recommended that the cause of this be investigated.