Eilmer3: Added 3D-turb-injector example.
Added the Turbulent Injector test case to
the 3D examples folder. Includes files for
the 3D case, the 2D runup, interpolator and
extrapolator python scripts and figures for
the latest solution. Please contact me for
the /flow files for the latest solution, as
further simulation is required before
steady-state is reached.