nenzfr: fixing a bug introduced in rev 1815...
As a part of the clean-up of the nozzle.input.template I removed
code that set the flow_condition to either 'initial' or 'inflow'
according to the blockMarching flag. I thought this code was
redundant since e3march propagates the inflow across the domain.

It turns out the code isn't redundant as e3march only propagates
the inflow for run > 0 (i.e. not for the first run). nenzfr
should be back up and running properly now...