dlang/gas: Swapped out JSON config for Lua config... part 1/2
The Ideal_gas model can now be constructed based on a Lua
table and using LuaD as the binding to Lua. The JSON style
configuration has been removed.

The makefile in the gas area is now aware of the LuaD
source files.

There was some fiddling required with compile options for
the Lua static library to get it to play nicely with the
D runtime. The Lua static library needs to be compiled
with the flag "-fno-omit-frame-pointer". I tested if this would
have any impact on using the static library with Eilmer3.
It appears there are no problems.

TODO: In gasmodelutil.d the init_gas_model() function
still tries to parse JSON files. I will remove that in
my next commit.

TODO: Fix the gas area makefile so that the Lua static
library is built if it has not already been built.