dlang/gas: Added CEA-type viscosity calculation + expt with org-mode.
This commit adds the CEA form of viscosity calculation.

I also experimented with adding documentation notes in org-mode
format. There are a few wrinkles to sort but I'm generally happy
with the process. The notes live in a subdirectory called
"notes" and the idea is to have a .org file to accompany each
.d file. I decided not to place the notes directly beside the
.d files because I didn't want to clutter those directories.