cfpylib/gasdyn: small changes to cea2_gas.py and gas_flow.py
I added a raise RuntimeError to part of cea2_gas.py so the error
trapping in Pitot can handle it better.

in finite_wave_dp in gas_flow.py I put an absolute value command
in the section that will add more steps if needed.
and in steady_flow_with_area_change I made the code drop the
initial guesses if the input pressure is very small, as that was
crashing Pitot sometimes