eilmer3/wallcon: Added Justin Beri's wallcon work to repository...
... and integrated it with Eilmer3. Wallcon is a stand-alone
solver for heat conduction. It also provides an interface
for interacting with Eilmer3 so that coupled gas/solid
heat transfer problems can be simulated.

Since we are now using Justin's finite-volume wall solver,
I also removed Jared Clifford's implementation from the working
tree. Eilmer3 no longer links to the twc library, but rather
the wallcon library.

Note, there is no conjugate heat transfer example cases
in this commit. They will follow when we've had a chance
to clean them up.

Also, this commit sneaks in work I did in a separate
repository on changing the precision with which eilmer3
writes grids and flow solutions. The details are captured
in this commit log messge:
   --------------------------------------------------------------------------
   changeset:   1652:ca3349fbd07b
   user:        Rowan J. Gollan <r.gollan@uq.edu.au>
   date:        Thu Aug 21 16:54:50 2014 +1000
   files:       app/eilmer3/source/bc.cxx app/eilmer3/source/block_bgk.cxx
                app/eilmer3/source/block_io.cxx app/eilmer3/source/cell.cxx
                app/eilmer3/source/e3_block.py app/eilmer3/source/e3_flow.py
                app/eilmer3/source/e3_grid.py app/eilmer3/source/e3prep.py
   description:
      eilmer3: Write out all flow and grid values with 16 significant figures.
   ---------------------------------------------------------------------------