dlang/gas: Thermally perfect gas mix caloric EOS added.
The caloric equation of state for a mixture of
thermally perfect gases has been added. The iterative
process to find temperature uses Ridder's method.
The bracketing nature of the method means that it
is more robust in the region of disjoing polynomials
that can be a problem with the CEA curves. In the limited
testing I've done, Ridder's method converges nicely and
avoids much of the complicated code required
when doing Newton iterations.

Also added the beginnings of a 'physical_constants'
module.