dlang/gas: Fixed const-correctness in diffusion module.
So that const can be properly applied to the gas model
methods, I needed to fix up the working arrays I use
in the Wilke mixing rule implementation. Following
PJ's suggestion, I have stored the working arrays
as static members in the Wilke mising classes.

Note: not sure what the ramifications of static data
are for parallelism. Will need to review this code then.