Add execute permission for radmodel.py so that it may be run as a command.
It's used in the 2D/radiating-cylinder/Argon/MC case and caused that case to
fail on a completely fresh install.  It seems that most people chmod +x the file
in the e3bin directory and are then happy because that permission flag seems
to be preserved from one install to the next.