Eilmer4: sort of issues of parsing JSON for configuration.
Had to clean up a bit of the manual JSON generation in e3prep.py.

The test.py example is essentially the cone20 simple case from Eilmer3.
Note, however, that the underlying gas model code has changed from C++ to D.
We'll need to think about how to handle that properly in time.  For the moment,
the test script initially uses the C++ model and then copies a replacement Lua
file into place.  This allows us to proceed with code development.