lib/nm: Alpha-QSS values can be set in an override file if desired.
There are a few tolerance parameters that affect the numerical
behaviour of the alpha-QSS chemistry integrator. The default
values I selected came from the combustion literature, however
they are not necessarily good values for doing high-temperature
air chemistry. As such, I wanted a way to set those values easily.
Previously, the defaults were hard-coded deep in the ODE initialisation
routines.

So to make those values settable from userland, I put in an ugly
hack where the code constructor will look for a specially named file
'override-alpha-qss-default-values.txt'. If the file exists,
the code will read the numeric parameters from that file. The file has
a fixed format that is documented by reading the code. The reason I
chose this ugly hack is that this code is now in maintenance mode
as we move to eilmer4. I did not feel like spending the effort of
updating the input code. It would have a ripple effect through to the
chemistry input and would require extra documentation in the user guides.