Eilmer4: implemented copy_into_ghost_cells for the FullFaceExchangeBC.
Should be able to reuse this function to exchange cell-geometry data when
setting up a new simulation.

Also, small edit to the ghost-cell copy in Eilmer3, just to keep the pattern.
It should make no difference to the simulation since bp.nnj and bp_src.nnj
have equal values for this orientation.