onedval: Added a 'brute' force option to increase robustness of optimiser.
Mat Bricalli had a set of troublesome data slices that presented problems
for the scipy Nelder-Mead optimiser. In the cases where the Nelder-Mead
optimiser consistently gives values outside of the valid range, the
scipy 'brute' force optimiser is used to search over a fixed grid of
values. This seems to do reasonably well for these duct flows.