Eilmer4: The core solver now steps without complaint.
The problem was that I had been ignoring the interface (grid) velocities which
I had thought irrelevant, and they would have been if they were zero.
The D-language compiler default initializes floating-point values to nan,
which is good but something that I had forgotten about.  These nans then
propagated when the flow velocities were transformed to grid-local values
and then fluxes calculated, and so on...

Anyway, we're now at the point where the solver steps in time but has yet to
do the cone20 simulation properly.