Eilmer3: read vertex moving velocity from C++ code
To speed up the simulation for fluid structure interaction, Eilmer3 now reads
vertex velocity from C++. However, there is now some problem with MPI, the
structural code should be executed as a whole block, but each block is simulated
individually with MPI. This results in "missing" vertex velocity data when
reading its data sometimes. At the current stage, when I did is to re-executed
the external structural code if the vertex velocities is missing. This should
be solved in the future.