Eilmer4: added -profile flag for instrumenting the code with dmd.
The output file from the instrumented run is found in trace.log.
The interesting stats on which functions take the processor time
may be a long way from the top of the file.
They seem to be starting at line 1832 this morning.

Remember to throw away this trace file before profiling the changed code,
else new stats will be accumulated on top of old.