Eilmer4: avoid invoking redundant copies of double arrays in fluxcalc.d.
This drops the number of calls to object._dup! for double[] to 77,000
down from 1,737,000 for the 50-step test.  Redundant copying of arrays
is no longer the nost expensive activity and the run times for the 840-step
test have reduced to:
dmd 2.066.1        -- 22 seconds
ldmd2 0.15.1       -- 21 seconds
Compared with:
e3shared g++ 4.8.2 -- 17 seconds
Given that Eilmer4 is doing a bit more copying and house-keeping with the
boundary-cell exchange, this is a good result.