dlang/gas: Removed @nogc, tidied file names, put gas module back in order.
After the brief dalliance with the @nogc attribute marker,
PJ and I decided that it is too restrictive and too pervasive
to be of use to us. Instead, we're going to rely on the tried
and true method of --ahem-- "disciplined" coding to avoid hidden
memory allocations.

While I was doing this, I dropped the use of thermalEOS and caloricEOS
in favour of pvt_eos and evt_eos. The former phrases seem antiquated
and caused confusion. I also made sure that file names had no capital
letters just to future-proof us against a dumb filesystem/OS that doesn't
honour case in file names. (There might be some files in the diffusion area
that require this change also.)