Eilmer4: finish the purge of code that makes redundant copies of double arrays.
This drops the number of calls to object._dup! for double[] to 44,360
down from 77,000 for the 50-step test in rev 1972.  Given that this is
a case with a 30x10 block and a 30x40 block, this number of calls to
make double arrays is close to the minimum to get the simulation set up.
Redundant copying of arrays is now in the noise.  There are, no doubt,
other places to make savings of computer effort.

Run times for the 840-step test are now:
dmd 2.066.1        -- 22 seconds
ldmd2 0.15.1       -- 22 seconds