dlang/geom: Added to the Lua-D wrapping experiment.
In this version, a global list of objects doesn't
need to be stored. It's equally clumsy, internally.
I actually think part of the problem is LuaD wrapped
up LuaState. I think some of this could be done neater
using the lua_state directly. The lua_state is available
as part of the LuaState class. I think I'll do that
in the next experiment.