dlang/geom: New attempt to wrap Vector3 struct for use in Lua scripts.
This attempt requires a bit of boilerplate code, but it
seems understandable, and robust. I'm committing this
now before I fiddle too much and break what's working.

Presently, I can access the x coordinate using a method syntax.
Assuming a = Vector3:new(), then
a:x(5.2) is a setter.
a:x() is a getter.
I would like to support the following syntax:
a.x = 5.2 is a setter.
a.x as a getter.