Eilmer4: structured-grid generation for surfaces is ported from e3prep.py
Also:
Renamed the clusterfunctions module to univariatefunctions because I moved
the implementation back to the generic classes seen in the C++ version.
It's more verbose than the Dlang  delegate implementation but
it gives me convenient places to put things like the string representation
function and it also forces a more explicit (and clearer) declaration
of the cluster-function parameter in the grid-gen functions.