dlang/geom: Minor tweak to luagpath.d to use references rather than pointers.
I realised that because the Paths are classes, I could use
the reference semantics to cut down on the use of pointers
in the codes. As a contrast, the wrapped Vector3 code uses pointers to
avoid the implicit copying that occurs with structs and their
value semantics. We need pointers rather than copies so
that we can change the value of the Vector3 for some
operations (eg. normalize).