Eilmer4: Added a Lua wrapper for FlowState.
This wrapped FlowState passes very minimal tests.
I have not actually tried to break its constructor.
After construction, users can get a view of the values
by asking for a Lua table version of the data. They
can't actually manipulate the data in the FlowState
once it is constructed[, presently --- need to think
if we'll allow that option.]