Eilmer4: migrated the initial flow file writing to D.
This was something that we wanted to do eventually but was done now to avoid
a segmentation fault when writing the files from Lua. The fault may have been
brought on by my misuse of Rowan's Lua wrapping functions with me returning
references rather than values for Vector3 structs in the StructuredGrid
indexing function.