Eilmer4: removed the redundant Lua functions for writing the initial flow files.
This commit done separately, just to keep a record.

Also, the current arrangement with gdata in the Lua input script and the
separate GlobalConfig not really being correctly updated works for the cone20
case only because the default values in GlobalConfig are OK for the initial
flow field writing.  Before going much further GlobalConfig has to be properly
updated during the running of the input script.