dlang/kinetics: Beginnings of the chemistry preparation program.
This work is largely a port of the chemistry preparation program
in lib/gas/kinetics with some tidy up. There is also a new strategy
for the preparation stage. In the old code, the user's chemistry
input was translated internally and passed directly to C++. In this
new code, the user-friendly Lua input is translated into a verbose Lua input.
The two formats are NOT interchangeable. The user is provided with convenience
functions and works in CGS units. The Lua input for the D code is a series
of verbose tables with all defaults filled in, and in SI units. It is meant
to be machine-readable from the D conde.  In this sense, the D code input
is no different from a JSON file had that been chosen as the input format.