dlang/kinetics: RKF update for chemistry has been verified as 5th order accurate.
The verification process with results is documented in:
dlang/kinetics/notes/rkf-verification.org

At last I've found a good use for literate programming. Org-mode
with the Babel package makes it reasonably easy to
document and build small programs and capture the results
of those programs in the one text file.