Eilmer4: Added the forward-facing step example.
There are a few code changes:
(1) Added more properties to the Lua table representation of a FlowState.
    (Wanted sound speed to make setting of the Mach 3 condition easier.)
(2) Tried to tighten up on the construction of SBlock objects in prep.lua.
(3) A couple of changes in luagrid and sgrid were added to rule out possible
    excuses for the preparation program segfaulting.

At this point, the new example still displays a segfault, sigh...