util: Remove failing test that tries to grabs floats as integers.
PJ and I keep getting different results when we get the D
type system to coerce a double into an int. For example,
on my machine, -15.8 converts to -15, on PJ's, it converts
to -16. It seems it's not portable to rely on coercing
a list of doubles into integer values... so we just won't do
that. I can't really see a need for this functionality, in any case.