eilmer4: Added n90 example, but non-reacting.
I've added the n90 example but it does not treat
the chemical reactions yet which cause dissociation
to occur. As such the shock wants to sit further
in front of the body than the present grid allows.
I won't change the grid because we will make this
case reacting very soon.

Along the way I fixed a bug in my lua_service function
getArrayOfNumbers. The values vector was not set
to "out" (ie. a ref) so the values is accumulated
were not reflected in the caller.

I also had to fill in some unimplemented functions
in the gas model so that sound speed was computed
for a thermally perfect gas.