Eilmer3: Update the Menter omega correction
Update the omega correction is applied for turbulent k-omega simulations,
following the recommendations in Menter's paper. Note that we now no longer
propagate the correction of the omega data in the first 6 cells (as we
previously did), since this is not recommended in Menter's paper.

This update seems to have alleviated the issues that we were seeing with
turbulent k-omega simulations failing due to low temperatures when very
cold walls are used (at least for the two cases that I have tested).

We will leave our previous implementation and notes of the Menter omega
correction in the code, just in case. The previous implementation should
be removed in a couple of months, should we not have any issues with the
updated version of the code.