Eilmer4: Added a basic reactor example.
This reactor is my old nitrogen test case.
The temperature is such that monatomic nitrogen
recombines to form diatomic nitrogen.
The equilibrium value looks good. I need to check
the transient when we have history output.

Also, I need to prettify this example. I've just
thrown the barebones files in as a placeholder.