Eilmer4: User-defined ghost cell bc implemented.
This commit adds the implementation of a UserDefinedGhostCell
boundary condition. I've also ported the cone20-udf example
and have it working for a user-defined supersonic inflow.
I need to build the sample_flow_data convenience function
before I can provide the other b.c.'s as user-defined
boundary conditions.

Note, the simulation code seg faults at the end of the simulation.
Presumably, this is happening during cleanup. There are two
references held to the LuaState. My guess is this might be the
problem. I will work on fixing this next.